xr-classifier/requirements.txt
fastapi>=0.111,<0.133
uvicorn[standard]>=0.30
pillow>=10.0,<11.0
numpy>=1.26,<2.0
scikit-learn>=1.4,<2.0
albumentation>=0.14
opencv-python-headless>=4.9
torch>=2.0,<3.0
torchvision>=0.15,<1.0
PyYAML>=0.46,<1.0